Refatora exercícios e adiciona exemplos de estruturas de dados

+++ Módulo 1 - Fundamentos/Aula 4 - Funções/desafio02.py
@@ -9,7 +9,7 @@ print("(1) Fahrenheit -> Celsius")
 print("(2) Celsius -> Fahrenheit")
 opcao = int(input("Sua opção:"))
 
-temp = int(input("Digite a temperatura:"))
+temp = float(input("Digite a temperatura:"))
 
 if opcao == 1:
     conversao = celsius(temp)
@@ -19,4 +19,3 @@ elif opcao == 2:
     print(f"A temperatura em fahrenheit é: {conversao:.1f}°F")
 else:
     print("Opcão Inválida!")
-
